+++ feedback_api/src/main/java/feedback_flow/feedback_api/application/configurations/AmazonSNSConfiguration.java
@@ -1,4 +1,4 @@
-package feedback_flow.feedback_api.domain.configuration;
+package feedback_flow.feedback_api.application.configurations;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
