+++ feedback_api/src/main/java/feedback_flow/feedback_api/application/repositories/CustomerFeedbackRepository.java
@@ -1,8 +1,8 @@
-package feedback_flow.feedback_api.domain.repository;
+package feedback_flow.feedback_api.application.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import feedback_flow.feedback_api.domain.model.CustomerFeedback;
+import feedback_flow.feedback_api.domain.customerFeedback.CustomerFeedback;
 
 public interface CustomerFeedbackRepository extends JpaRepository<CustomerFeedback, Long> {
 }
