+++ feedback_api/src/main/java/feedback_flow/feedback_api/application/sevices/CustomerFeedbackService.java
@@ -1,4 +1,4 @@
-package feedback_flow.feedback_api.domain.service;
+package feedback_flow.feedback_api.application.sevices;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -6,16 +6,16 @@ import org.springframework.stereotype.Service;
 import com.amazonaws.services.sns.AmazonSNS;
 import com.amazonaws.services.sns.model.PublishResult;
 
-import feedback_flow.feedback_api.domain.configuration.FeedbackTopicMapping;
-import feedback_flow.feedback_api.domain.model.CustomerFeedbackType;
+import feedback_flow.feedback_api.application.configurations.FeedbackTopicMappingConfiguration;
+import feedback_flow.feedback_api.domain.customerFeedback.CustomerFeedbackType;
 
 @Service
 public class CustomerFeedbackService extends AmazonSNSService {
 
-    private final FeedbackTopicMapping topicMapping;
+    private final FeedbackTopicMappingConfiguration topicMapping;
 
     @Autowired
-    public CustomerFeedbackService(AmazonSNS amazonSNS, FeedbackTopicMapping topicMapping) {
+    public CustomerFeedbackService(AmazonSNS amazonSNS, FeedbackTopicMappingConfiguration topicMapping) {
         super(amazonSNS);
         this.topicMapping = topicMapping;
     }
