+++ feedback_api/src/main/java/feedback_flow/feedback_api/domain/customerFeedback/CustomerFeedback.java
@@ -1,4 +1,4 @@
-package feedback_flow.feedback_api.domain.model;
+package feedback_flow.feedback_api.domain.customerFeedback;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
