+++ feedback_api/src/main/java/feedback_flow/feedback_api/domain/customerFeedback/CustomerFeedbackType.java
@@ -1,4 +1,4 @@
-package feedback_flow.feedback_api.domain.model;
+package feedback_flow.feedback_api.domain.customerFeedback;
 
 public enum CustomerFeedbackType {
     SUGGESTION,
